MySQL: use INFORMATION_SCHEMA.ST_GEOMETRY_COLUMNS to get SRSId

+++ gdal/ogr/ogrsf_frmts/mysql/ogrmysqldatasource.cpp
@@ -387,6 +387,10 @@ OGRLayer *OGRMySQLDataSource::GetLayer( int iLayer )
         return papoLayers[iLayer];
 }
 
+/* =====================================================================*/
+/* Handle spatial reference id and index for older MySQL and MariaDB    */
+/* =====================================================================*/
+
 /************************************************************************/
 /*                      InitializeMetadataTables()                      */
 /*                                                                      */
@@ -402,37 +406,37 @@ OGRErr OGRMySQLDataSource::InitializeMetadataTables()
     MYSQL_RES       *hResult;
     OGRErr          eErr = OGRERR_NONE;
 
-    pszCommand = "DESCRIBE geometry_columns";
-    if( mysql_query(GetConn(), pszCommand ) )
-    {
-        pszCommand =
-                "CREATE TABLE geometry_columns "
-                "( F_TABLE_CATALOG VARCHAR(256), "
-                "F_TABLE_SCHEMA VARCHAR(256), "
-                "F_TABLE_NAME VARCHAR(256) NOT NULL,"
-                "F_GEOMETRY_COLUMN VARCHAR(256) NOT NULL, "
-                "COORD_DIMENSION INT, "
-                "SRID INT,"
-                "TYPE VARCHAR(256) NOT NULL)";
-        if( mysql_query(GetConn(), pszCommand ) )
-        {
-            ReportError( pszCommand );
-            eErr = OGRERR_FAILURE;
-        }
-        else
-            CPLDebug("MYSQL","Creating geometry_columns metadata table");
-    }
-
-    // make sure to attempt to free results of successful queries
-    hResult = mysql_store_result( GetConn() );
-    if( hResult != nullptr )
-    {
-        mysql_free_result( hResult );
-        hResult = nullptr;
-    }
-
     if( GetMajorVersion() < 8 || IsMariaDB() )
     {
+        pszCommand = "DESCRIBE geometry_columns";
+        if( mysql_query(GetConn(), pszCommand ) )
+        {
+            pszCommand =
+                    "CREATE TABLE geometry_columns "
+                    "( F_TABLE_CATALOG VARCHAR(256), "
+                    "F_TABLE_SCHEMA VARCHAR(256), "
+                    "F_TABLE_NAME VARCHAR(256) NOT NULL,"
+                    "F_GEOMETRY_COLUMN VARCHAR(256) NOT NULL, "
+                    "COORD_DIMENSION INT, "
+                    "SRID INT,"
+                    "TYPE VARCHAR(256) NOT NULL)";
+            if( mysql_query(GetConn(), pszCommand ) )
+            {
+                ReportError( pszCommand );
+                eErr = OGRERR_FAILURE;
+            }
+            else
+                CPLDebug("MYSQL","Creating geometry_columns metadata table");
+        }
+
+        // make sure to attempt to free results of successful queries
+        hResult = mysql_store_result( GetConn() );
+        if( hResult != nullptr )
+        {
+            mysql_free_result( hResult );
+            hResult = nullptr;
+        }
+
         pszCommand = "DESCRIBE spatial_ref_sys";
         if( mysql_query(GetConn(), pszCommand ) )
         {
@@ -462,6 +466,8 @@ OGRErr OGRMySQLDataSource::InitializeMetadataTables()
 
     return eErr;
 }
+
+/* =====================================================================*/
 
 /************************************************************************/
 /*                              FetchSRS()                              */
@@ -1099,64 +1105,21 @@ OGRMySQLDataSource::ICreateLayer( const char * pszLayerNameIn,
         mysql_free_result( hResult );
     hResult = nullptr;
 
+/* =====================================================================*/
+/*  Handle spatial id and index separately for older MySQL and MariaDB  */
+/* -------------------------------------------------------------------- */
+
+    if( GetMajorVersion() < 8 || IsMariaDB() )
+    {
 /* -------------------------------------------------------------------- */
 /*      Sometimes there is an old crufty entry in the geometry_columns  */
 /*      table if things were not properly cleaned up before.  We make   */
 /*      an effort to clean out such cruft.                              */
 /*                                                                      */
 /* -------------------------------------------------------------------- */
-    osCommand.Printf(
-             "DELETE FROM geometry_columns WHERE f_table_name = '%s'",
-             pszLayerName );
-
-    if( mysql_query(GetConn(), osCommand ) )
-    {
-        ReportError( osCommand );
-        return nullptr;
-    }
-
-    // make sure to attempt to free results of successful queries
-    hResult = mysql_store_result( GetConn() );
-    if( hResult != nullptr )
-        mysql_free_result( hResult );
-    hResult = nullptr;
-
-/* -------------------------------------------------------------------- */
-/*      Attempt to add this table to the geometry_columns table, if     */
-/*      it is a spatial layer.                                          */
-/* -------------------------------------------------------------------- */
-    if( eType != wkbNone )
-    {
-        const int nCoordDimension = eType == wkbFlatten(eType) ? 2 : 3;
-
-        pszGeometryType = OGRToOGCGeomType(eType);
-
-        if( nSRSId == GetUnknownSRID() )
-            osCommand.Printf(
-                     "INSERT INTO geometry_columns "
-                     " (F_TABLE_NAME, "
-                     "  F_GEOMETRY_COLUMN, "
-                     "  COORD_DIMENSION, "
-                     "  TYPE) values "
-                     "  ('%s', '%s', %d, '%s')",
-                     pszLayerName,
-                     pszGeomColumnName,
-                     nCoordDimension,
-                     pszGeometryType );
-        else
-            osCommand.Printf(
-                     "INSERT INTO geometry_columns "
-                     " (F_TABLE_NAME, "
-                     "  F_GEOMETRY_COLUMN, "
-                     "  COORD_DIMENSION, "
-                     "  SRID, "
-                     "  TYPE) values "
-                     "  ('%s', '%s', %d, %d, '%s')",
-                     pszLayerName,
-                     pszGeomColumnName,
-                     nCoordDimension,
-                     nSRSId,
-                     pszGeometryType );
+        osCommand.Printf(
+                 "DELETE FROM geometry_columns WHERE f_table_name = '%s'",
+                 pszLayerName );
 
         if( mysql_query(GetConn(), osCommand ) )
         {
@@ -1169,7 +1132,58 @@ OGRMySQLDataSource::ICreateLayer( const char * pszLayerNameIn,
         if( hResult != nullptr )
             mysql_free_result( hResult );
         hResult = nullptr;
+
+/* -------------------------------------------------------------------- */
+/*      Attempt to add this table to the geometry_columns table, if     */
+/*      it is a spatial layer.                                          */
+/* -------------------------------------------------------------------- */
+        if( eType != wkbNone )
+        {
+            const int nCoordDimension = eType == wkbFlatten(eType) ? 2 : 3;
+
+            pszGeometryType = OGRToOGCGeomType(eType);
+
+            if( nSRSId == GetUnknownSRID() )
+                osCommand.Printf(
+                         "INSERT INTO geometry_columns "
+                         " (F_TABLE_NAME, "
+                         "  F_GEOMETRY_COLUMN, "
+                         "  COORD_DIMENSION, "
+                         "  TYPE) values "
+                         "  ('%s', '%s', %d, '%s')",
+                         pszLayerName,
+                         pszGeomColumnName,
+                         nCoordDimension,
+                         pszGeometryType );
+            else
+                osCommand.Printf(
+                         "INSERT INTO geometry_columns "
+                         " (F_TABLE_NAME, "
+                         "  F_GEOMETRY_COLUMN, "
+                         "  COORD_DIMENSION, "
+                         "  SRID, "
+                         "  TYPE) values "
+                         "  ('%s', '%s', %d, %d, '%s')",
+                         pszLayerName,
+                         pszGeomColumnName,
+                         nCoordDimension,
+                         nSRSId,
+                         pszGeometryType );
+
+            if( mysql_query(GetConn(), osCommand ) )
+            {
+                ReportError( osCommand );
+                return nullptr;
+            }
+
+            // make sure to attempt to free results of successful queries
+            hResult = mysql_store_result( GetConn() );
+            if( hResult != nullptr )
+                mysql_free_result( hResult );
+            hResult = nullptr;
+        }
     }
+/* =====================================================================*/
 
 /* -------------------------------------------------------------------- */
 /*      Create the spatial index.                                       */
